Skilt kontonummer-form ut i egen komponent, fjernet overflødige tags i telefonnummer-form
